types: Refactor 'debug' module into 'types'
