--- module/debug/tsconfig.json
@@ -1,14 +1,0 @@
-{
-	"extends": "../../tsconfig.settings.json",
-	"compilerOptions": {
-		"baseUrl": ".",
-		"outDir": "scripts",
-		"declarationDir": "types"
-	},
-	"include": [
-		"src/**/*.ts"
-	],
-	"tests": [
-		"test/**/*.ts"
-	]
-}
