+++ module/types/src/Uint8.ts
@@ -5,8 +5,7 @@
 //. Developer Notes:
 //. @eth-p: We could use Uint8Array to do the conversions, but it was actually slower in every browser except Chrome.
 //. --------------------------------------------------------------------------------------------------------------------
-import assert from '@chipotle/debug/assert';
-
+import assert from './assert';
 import MathFlag from './MathFlag';
 import MathResult from './MathResult';
 // ---------------------------------------------------------------------------------------------------------------------
