+++ tsconfig.settings.json
@@ -20,7 +20,6 @@
 		"noEmit": true,
 		"paths": {
 			"@chipotle/types/*": ["../types/src/*"],
-			"@chipotle/debug/*": ["../debug/src/*"],
 			"@chipotle/vm/*": ["../vm/src/*"],
 			"@chipotle/arch-chip/*": ["../arch-chip/src/*"]
 		}
